requirements.txt
patchright>=0.9.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyYAML>=6.0.1
click>=8.1.7
structlog>=24.1.0
